rename source code folder to pyeasyfuse
