+++ .github/workflows/unit_tests.yml
@@ -18,6 +18,6 @@ jobs:
     - name: Build, install and run unit tests
       run: |
         poetry install
-        poetry run coverage run -m unittest discover easy_fuse.tests
+        poetry run coverage run -m unittest discover pyeasyfuse.tests
     - name: Upload Coverage to Codecov
       uses: codecov/codecov-action@v3
