+++ pyeasyfuse/tests/resources/fusioncatcher_1.txt
