+++ pyeasyfuse/tests/resources/fusioncatcher_2.txt
