+++ pyeasyfuse/tests/resources/infusion.txt
