+++ pyeasyfuse/tests/resources/mapsplice.txt
