+++ pyeasyfuse/tests/resources/soapfuse.txt
