+++ pyeasyfuse/tests/resources/starfusion.txt
